feat #29: 타겟 분리
