--- soonwook/Book/Data/Entity/BookResponse.swift
@@ -1,21 +1,0 @@
-//
-//  BookResponse.swift
-//  Book
-//
-//  Created by 권순욱 on 5/14/25.
-//
-
-import Foundation
-
-struct BookResponse: Decodable {
-    let documents: [Book]
-    let meta: Meta
-}
-
-struct Meta: Decodable {
-    let isEnd: Bool
-    
-    enum CodingKeys: String, CodingKey {
-        case isEnd = "is_end"
-    }
-}
