--- soonwook/Book/Data/Entity/CartItem.swift
@@ -1,15 +1,0 @@
-//
-//  CartItem.swift
-//  Book
-//
-//  Created by 권순욱 on 5/13/25.
-//
-
-import Foundation
-
-struct CartItem {
-    let isbn: String
-    let title: String
-    let author: String
-    let price: Int
-}
