+++ soonwook/DomainLayer/DomainLayer.h
@@ -1,0 +1,18 @@
+//
+//  DomainLayer.h
+//  DomainLayer
+//
+//  Created by 권순욱 on 5/15/25.
+//
+
+#import <Foundation/Foundation.h>
+
+//! Project version number for DomainLayer.
+FOUNDATION_EXPORT double DomainLayerVersionNumber;
+
+//! Project version string for DomainLayer.
+FOUNDATION_EXPORT const unsigned char DomainLayerVersionString[];
+
+// In this header, you should import all the public headers of your framework using statements like #import <DomainLayer/PublicHeader.h>
+
+
