+++ soonwook/DomainLayer/Model/CartItem.swift
@@ -1,0 +1,15 @@
+//
+//  CartItem.swift
+//  Book
+//
+//  Created by 권순욱 on 5/13/25.
+//
+
+import Foundation
+
+public struct CartItem {
+    public let isbn: String
+    public let title: String
+    public let author: String
+    public let price: Int
+}
